agent-python/pyproject.toml
[project]
name = "agent-python"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[cartesia,deepgram,groq,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
